+++ packages/react-native-leader-line/jest.config.edge.js
@@ -1,0 +1,5 @@
+module.exports = {
+  ...require('./jest.config.js'),
+  testEnvironment: 'jsdom',
+  testMatch: ['<rootDir>/src/components/__tests__/edge-cases.test.tsx'],
+};