+++ packages/react-native-leader-line/jest.config.js
@@ -1,11 +1,12 @@
 module.exports = {
-  preset: 'react-native',
+
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
   setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
   testMatch: [
     '<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}',
     '<rootDir>/src/**/*.{test,spec}.{js,jsx,ts,tsx}'
   ],
+  testPathIgnorePatterns: ['<rootDir>/src/components/__tests__/edge-cases.test.tsx'],
   collectCoverageFrom: [
     'src/**/*.{js,jsx,ts,tsx}',
     '!src/**/*.d.ts',
@@ -13,7 +14,7 @@ module.exports = {
   ],
   coverageDirectory: 'coverage',
   transform: {
-    '^.+\\.(js|jsx|ts|tsx)$': 'babel-jest',
+    '^.+\.(js|jsx|ts|tsx)$': 'babel-jest',
   },
   transformIgnorePatterns: [
     'node_modules/(?!(react-native|@react-native|react-native-svg|@testing-library)/)'
@@ -21,7 +22,8 @@ module.exports = {
   moduleNameMapper: {
     '^react-native-svg$': '<rootDir>/__mocks__/react-native-svg.js'
   },
-  testEnvironment: 'node',
+
+  preset: 'react-native',
   globals: {
     __DEV__: true,
   },